tool/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
reportlab==4.1.0
openpyxl==3.1.2
pandas==2.2.2
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dateutil==2.9.0
apscheduler==3.10.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
